+++ ginga/rv/plugins/Header.py
@@ -27,53 +27,90 @@ was created with an option not to save the primary header.
 from collections import OrderedDict
 
 from ginga import GingaPlugin
-from ginga.misc import Bunch
 from ginga.gw import Widgets
 
 __all__ = ['Header']
 
 
-class Header(GingaPlugin.GlobalPlugin):
+class Header(GingaPlugin.ParentPlugin):
+    # Header uses the ParentPlugin class to share the same window amongst
+    # a bunch of local plugins
 
     def __init__(self, fv):
         # superclass defines some variables for us, like logger
-        super(Header, self).__init__(fv)
+        super().__init__(fv)
 
-        self._image = None
-        self.active = None
-        self.info = None
+        # the child plugin class we will start for each channel opened
+        self.class_childplugin = HeaderChild
+
+        spec = self.fv.get_plugin_spec(str(self))
+
+        prefs = self.fv.get_preferences()
+        self.settings = prefs.create_category('plugin_Header')
+        self.settings.add_defaults(closeable=not spec.get('hidden', False))
+        self.settings.load(onError='silent')
+
+        self.gui_up = False
+
+    def build_gui(self, container):
+        super().build_gui(container)
+
+        if self.settings.get('closeable', False):
+            btns = Widgets.HBox()
+            btns.set_border_width(4)
+            btns.set_spacing(4)
+
+            btn = Widgets.Button("Close")
+            btn.add_callback('activated', lambda w: self.close())
+            btns.add_widget(btn)
+            btn = Widgets.Button("Help")
+            btn.add_callback('activated', lambda w: self.help())
+            btns.add_widget(btn, stretch=0)
+            btns.add_widget(Widgets.Label(''), stretch=1)
+
+            container.add_widget(btns, stretch=0)
+        self.gui_up = True
+
+    def __str__(self):
+        return 'header'
+
+
+class HeaderChild(GingaPlugin.ChildPlugin):
+    # This local plugin actually does all the work, but only has to be
+    # concerned with tracking a single channel
+
+    def __init__(self, fv, image_viewer, parent_plugin):
+        # superclass defines some variables for us, like logger
+        super().__init__(fv, image_viewer, parent_plugin)
+
         self.columns = [('Keyword', 'key'),
                         ('Value', 'value'),
                         ('Comment', 'comment'),
                         ]
-
-        spec = self.fv.get_plugin_spec(str(self))
 
         prefs = self.fv.get_preferences()
         self.settings = prefs.create_category('plugin_Header')
         self.settings.add_defaults(sortable=False,
                                    color_alternate_rows=True,
                                    max_rows_for_col_resize=5000,
-                                   include_primary_header=False,
-                                   closeable=not spec.get('hidden', False))
+                                   include_primary_header=False)
         self.settings.load(onError='silent')
 
         self.flg_sort = self.settings.get('sortable', False)
         self.flg_prihdr = self.settings.get('include_primary_header', False)
-        fv.add_callback('add-channel', self.add_channel)
-        fv.add_callback('delete-channel', self.delete_channel)
-        fv.add_callback('channel-change', self.focus_cb)
-
-        self.gui_up = False
 
     def build_gui(self, container):
         vbox = Widgets.VBox()
-        vbox.set_border_width(1)
+        vbox.set_border_width(2)
         vbox.set_spacing(1)
 
-        nb = Widgets.StackWidget()
-        vbox.add_widget(nb, stretch=1)
-        self.nb = nb
+        color_alternate = self.settings.get('color_alternate_rows', True)
+        table = Widgets.TreeView(auto_expand=True,
+                                 use_alt_row_color=color_alternate)
+        self.w.table = table
+        table.setup_table(self.columns, 1, 'key')
+
+        vbox.add_widget(table, stretch=1)
 
         # create sort toggle
         hbox = Widgets.HBox()
@@ -89,39 +126,11 @@ class Header(GingaPlugin.GlobalPlugin):
         hbox.add_widget(Widgets.Label(''), stretch=1)
         vbox.add_widget(hbox, stretch=0)
 
-        if self.settings.get('closeable', False):
-            btns = Widgets.HBox()
-            btns.set_border_width(4)
-            btns.set_spacing(4)
-
-            btn = Widgets.Button("Close")
-            btn.add_callback('activated', lambda w: self.close())
-            btns.add_widget(btn)
-            btn = Widgets.Button("Help")
-            btn.add_callback('activated', lambda w: self.help())
-            btns.add_widget(btn, stretch=0)
-            btns.add_widget(Widgets.Label(''), stretch=1)
-            vbox.add_widget(btns, stretch=0)
-
         container.add_widget(vbox, stretch=1)
+
         self.gui_up = True
 
-    def _create_header_window(self, info):
-        vbox = Widgets.VBox()
-        vbox.set_margins(2, 2, 2, 2)
-
-        color_alternate = self.settings.get('color_alternate_rows', True)
-        table = Widgets.TreeView(auto_expand=True,
-                                 use_alt_row_color=color_alternate)
-        self.table = table
-        table.setup_table(self.columns, 1, 'key')
-
-        vbox.add_widget(table, stretch=1)
-
-        info.setvals(widget=vbox, table=table)
-        return vbox
-
-    def set_header(self, info, image):
+    def set_header(self, image):
         if self._image == image:
             # we've already handled this header
             return
@@ -133,7 +142,7 @@ class Header(GingaPlugin.GlobalPlugin):
             self.w.chk_prihdr.set_enabled(has_prihdr)
 
         header = image.get_header(include_primary_header=self.flg_prihdr)
-        table = info.table
+        table = self.w.table
 
         is_sorted = self.flg_sort
         tree_dict = OrderedDict()
@@ -156,113 +165,37 @@ class Header(GingaPlugin.GlobalPlugin):
         self.logger.debug("setting header done ({0})".format(is_sorted))
         self._image = image
 
-    def add_channel(self, viewer, channel):
-        if not self.gui_up:
-            return
-        chname = channel.name
-        info = Bunch.Bunch(chname=chname)
-        sw = self._create_header_window(info)
-
-        self.nb.add_widget(sw)
-        info.setvals(widget=sw)
-        channel.extdata._header_info = info
-
-    def delete_channel(self, viewer, channel):
-        if not self.gui_up:
-            return
-        chname = channel.name
-        self.logger.debug("deleting channel %s" % (chname))
-        info = channel.extdata._header_info
-        widget = info.widget
-        self.nb.remove(widget, delete=True)
-        self.active = None
-        self.info = None
-
-    def focus_cb(self, viewer, channel):
-        if not self.gui_up:
-            return
-        chname = channel.name
-
-        if self.active != chname:
-            if '_header_info' not in channel.extdata:
-                self.add_channel(viewer, channel)
-            info = channel.extdata._header_info
-            widget = info.widget
-            index = self.nb.index_of(widget)
-            self.nb.set_index(index)
-            self.active = chname
-            self.info = info
-
-        image = channel.get_current_image()
-        if image is None:
-            return
-        self.set_header(self.info, image)
-
     def start(self):
-        names = self.fv.get_channel_names()
-        for name in names:
-            channel = self.fv.get_channel(name)
-            self.add_channel(self.fv, channel)
-
-        channel = self.fv.get_channel_info()
-        if channel is not None:
-            viewer = channel.fitsimage
-
-            image = viewer.get_image()
-            if image is not None:
-                self.redo(channel, image)
-
-            self.focus_cb(viewer, channel)
+        pass
 
     def stop(self):
-        names = self.fv.get_channel_names()
-        for name in names:
-            channel = self.fv.get_channel(name)
-            channel.extdata._header_info = None
-
         self.gui_up = False
-        self.nb = None
-        self.active = None
-        self.info = None
+
+    def close(self):
+        #self.fv.stop_local_plugin(self.chname, str(self))
         return True
 
-    def redo(self, channel, image):
+    def redo(self):
         """This is called when image changes."""
         self._image = None  # Skip cache checking in set_header()
-        info = channel.extdata._header_info
 
-        self.set_header(info, image)
-
-    def blank(self, channel):
-        """This is called when image is cleared."""
-        self._image = None
-        info = channel.extdata._header_info
-        info.table.clear()
+        # TODO: what about tables and the table viewer?
+        image = self.fitsimage.get_image()
+        if image is None:
+            self.w.table.clear()
+        else:
+            self.set_header(image)
 
     def set_sortable_cb(self, tf):
         self.flg_sort = tf
-        self._image = None
-        if self.info is not None:
-            info = self.info
-            channel = self.fv.get_channel(info.chname)
-            image = channel.get_current_image()
-            self.set_header(info, image)
+        self.redo()
 
     def set_prihdr_cb(self, tf):
         self.flg_prihdr = tf
-        self._image = None
-        if self.info is not None:
-            info = self.info
-            channel = self.fv.get_channel(info.chname)
-            image = channel.get_current_image()
-            self.set_header(info, image)
-
-    def close(self):
-        self.fv.stop_global_plugin(str(self))
-        return True
+        self.redo()
 
     def __str__(self):
-        return 'header'
+        return 'headerchild'
 
 
 # Append module docstring with config doc for auto insert by Sphinx.