+++ ginga/rv/plugins/Pan.py
@@ -8,7 +8,7 @@ The ``Pan`` plugin provides a small panning image that gives an overall
 channel image is zoomed in 2X or greater, then the pan region is shown
 graphically in the ``Pan`` image by a rectangle.
 
-**Plugin Type: Global**
+**Plugin Type: Local**
 
 ``Pan`` is a global plugin.  Only one instance can be opened.
 
@@ -40,53 +40,32 @@ import traceback
 import math
 
 from ginga.gw import Widgets, Viewers
-from ginga.misc import Bunch
 from ginga.util import wcs
 from ginga import GingaPlugin
 
 __all__ = ['Pan']
 
 
-class Pan(GingaPlugin.GlobalPlugin):
+class Pan(GingaPlugin.ParentPlugin):
+    # Pan uses the ParentPlugin class to share the same window amongst
+    # a bunch of local plugins
 
     def __init__(self, fv):
-        # superclass defines some variables for us, like logger
-        super(Pan, self).__init__(fv)
+        super().__init__(fv)
 
-        self.active = None
-        self.info = None
-
-        fv.add_callback('add-channel', self.add_channel)
-        fv.add_callback('delete-channel', self.delete_channel)
-        fv.set_callback('channel-change', self.focus_cb)
-
-        self.dc = fv.get_draw_classes()
+        # the child plugin class we will start for each channel opened
+        self.class_childplugin = PanChild
 
         spec = self.fv.get_plugin_spec(str(self))
-
         prefs = self.fv.get_preferences()
         self.settings = prefs.create_category('plugin_Pan')
-        self.settings.add_defaults(closeable=not spec.get('hidden', False),
-                                   pan_position_color='yellow',
-                                   pan_rectangle_color='red',
-                                   compass_color='skyblue',
-                                   rotate_pan_image=True)
+        self.settings.add_defaults(closeable=not spec.get('hidden', False))
         self.settings.load(onError='silent')
 
-        self._wd = 200
-        self._ht = 200
-        self.copy_attrs = ['transforms', 'cutlevels', 'rotation', 'rgbmap',
-                           'limits', 'icc', 'interpolation']
         self.gui_up = False
 
     def build_gui(self, container):
-        vbox = Widgets.VBox()
-        vbox.set_border_width(2)
-        vbox.set_spacing(2)
-
-        nb = Widgets.StackWidget()
-        self.nb = nb
-        vbox.add_widget(self.nb, stretch=1)
+        super().build_gui(container)
 
         if self.settings.get('closeable', False):
             btns = Widgets.HBox()
@@ -100,12 +79,48 @@ class Pan(GingaPlugin.GlobalPlugin):
             btn.add_callback('activated', lambda w: self.help())
             btns.add_widget(btn, stretch=0)
             btns.add_widget(Widgets.Label(''), stretch=1)
-            vbox.add_widget(btns, stretch=0)
 
-        container.add_widget(vbox, stretch=1)
+            container.add_widget(btns, stretch=0)
         self.gui_up = True
 
-    def _create_pan_viewer(self, fitsimage):
+    def __str__(self):
+        return 'pan'
+
+
+class PanChild(GingaPlugin.ChildPlugin):
+    # This local plugin actually does all the work, but only has to be
+    # concerned with tracking a single channel
+
+    def __init__(self, fv, image_viewer, parent_plugin):
+        # superclass defines some variables for us, like logger
+        super().__init__(fv, image_viewer, parent_plugin)
+
+        self.dc = fv.get_draw_classes()
+
+        prefs = self.fv.get_preferences()
+        self.settings = prefs.create_category('plugin_Pan')
+        self.settings.add_defaults(pan_position_color='yellow',
+                                   pan_rectangle_color='red',
+                                   compass_color='skyblue',
+                                   rotate_pan_image=True)
+        self.settings.load(onError='silent')
+
+        self._wd = 200
+        self._ht = 200
+        self.copy_attrs = ['transforms', 'cutlevels', 'rotation', 'rgbmap',
+                           'limits', 'icc', 'interpolation']
+
+        self.compass_xy = None
+        self.compass_wcs = None
+        self.panrect = None
+
+        self.gui_up = False
+
+    def build_gui(self, container):
+        vbox = Widgets.VBox()
+        vbox.set_border_width(2)
+        vbox.set_spacing(2)
+
         pi = Viewers.CanvasView(logger=self.logger)
         pi.enable_autozoom('on')
         pi.enable_autocuts('off')
@@ -122,145 +137,75 @@ class Pan(GingaPlugin.GlobalPlugin):
         pi.set_callback('configure', self.reconfigure)
         # for debugging
         pi.set_name('panimage')
+        self.panimage = pi
 
         my_canvas = pi.get_private_canvas()
         my_canvas.enable_draw(True)
         my_canvas.set_drawtype('rectangle', linestyle='dash', color='green')
         my_canvas.set_callback('draw-event', self.draw_cb)
 
-        canvas = fitsimage.get_canvas()
+        # we share a canvas with the main channel image viewer
+        canvas = self.fitsimage.get_canvas()
         pi.set_canvas(canvas)
 
         bd = pi.get_bindings()
         bd.enable_pan(False)
         bd.enable_zoom(False)
 
-        return pi
-
-    def add_channel(self, viewer, channel):
-        if not self.gui_up:
-            return
-        fitsimage = channel.fitsimage
-        panimage = self._create_pan_viewer(fitsimage)
-        p_canvas = panimage.get_private_canvas()
+        p_canvas = pi.get_private_canvas()
 
         # add X/Y compass
         x, y = 0.5, 0.5
         radius = 0.1
 
-        compass_xy = self.dc.Compass(x, y, radius,
-                                     color=self.settings.get('xy_compass_color',
-                                                             'yellow'),
-                                     fontsize=14, ctype='pixel',
-                                     coord='percentage')
-        p_canvas.add_object(compass_xy)
+        self.compass_xy = self.dc.Compass(x, y, radius,
+                                          color=self.settings.get('xy_compass_color',
+                                                                  'yellow'),
+                                          fontsize=14, ctype='pixel',
+                                          coord='percentage')
+        p_canvas.add_object(self.compass_xy)
 
-        compass_wcs = self.dc.Compass(x, y, radius * 2.0,
-                                      color=self.settings.get('compass_color',
-                                                              'skyblue'),
-                                      fontsize=14, ctype='wcs',
-                                      coord='percentage')
+        self.compass_wcs = self.dc.Compass(x, y, radius * 2.0,
+                                           color=self.settings.get('compass_color',
+                                                                   'skyblue'),
+                                           fontsize=14, ctype='wcs',
+                                           coord='percentage')
 
-        iw = Viewers.GingaViewerWidget(panimage)
-        iw.resize(self._wd, self._ht)
-        self.nb.add_widget(iw)
-        paninfo = Bunch.Bunch(panimage=panimage, widget=iw,
-                              compass_wcs=compass_wcs, compass_xy=compass_xy,
-                              panrect=None)
-        channel.extdata._pan_info = paninfo
+        pw = Viewers.GingaViewerWidget(pi)
+        pw.resize(self._wd, self._ht)
 
-        fitsimage.copy_attributes(panimage, self.copy_attrs)
+        self.fitsimage.copy_attributes(pi, self.copy_attrs)
+        self.fitsimage.add_callback('redraw', self.redraw_cb)
 
-        fitsimage.add_callback('redraw', self.redraw_cb, channel)
-        self.logger.debug("channel '%s' added." % (channel.name))
-
-    def delete_channel(self, viewer, channel):
-        if not self.gui_up:
-            return
-        chname = channel.name
-        self.logger.debug("deleting channel %s" % (chname))
-        widget = channel.extdata._pan_info.widget
-        channel.extdata._pan_info.widget = None
-        self.nb.remove(widget, delete=True)
-        self.active = None
-        self.info = None
+        vbox.add_widget(pw, stretch=1)
+        container.add_widget(vbox, stretch=1)
+        self.gui_up = True
 
     def start(self):
-        names = self.fv.get_channel_names()
-        for name in names:
-            channel = self.fv.get_channel(name)
-            self.add_channel(self.fv, channel)
-
-        channel = self.fv.get_channel_info()
-        if channel is not None:
-            viewer = channel.fitsimage
-
-            image = viewer.get_image()
-            if image is not None:
-                self.redo(channel, image)
-
-            self.focus_cb(viewer, channel)
+        pass
 
     def stop(self):
-        names = self.fv.get_channel_names()
-        for name in names:
-            channel = self.fv.get_channel(name)
-            channel.extdata._pan_info = None
-
-        self.active = None
-        self.nb = None
-        self.info = None
         self.gui_up = False
-
-    def close(self):
-        self.fv.stop_global_plugin(str(self))
-        return True
 
     # CALLBACKS
 
-    def redo(self, channel, image):
+    def redo(self):
         if not self.gui_up:
             return
-        self.logger.debug("redo")
-        paninfo = channel.extdata._pan_info
 
-        if (image is None) or not paninfo.panimage.viewable(image):
+        image = self.fitsimage.get_image()
+
+        if (image is None) or not self.panimage.viewable(image):
             self.logger.debug("no main image--clearing Pan viewer")
-            paninfo.panimage.clear()
+            self.panimage.clear()
             return
 
-        self.set_image(channel, paninfo, image)
+        self.set_image(image)
 
     def blank(self, channel):
         if not self.gui_up:
             return
-        paninfo = channel.extdata._pan_info
-        paninfo.panimage.clear()
-
-    def focus_cb(self, viewer, channel):
-        if not self.gui_up:
-            return
-        chname = channel.name
-
-        # If the active widget has changed, then raise our Info widget
-        # that corresponds to it
-        if self.active != chname:
-            if '_pan_info' not in channel.extdata:
-                self.add_channel(viewer, channel)
-            paninfo = channel.extdata._pan_info
-            iw = paninfo.widget
-            index = self.nb.index_of(iw)
-            self.nb.set_index(index)
-            self.active = chname
-            self.info = paninfo
-
-        # TODO: this check should not be necessary.  But under some
-        # circumstances it seems to be needed.
-        image = channel.fitsimage.get_image()
-        p_image = self.info.panimage.get_image()
-        if image != p_image:
-            self.logger.debug("pan viewer seems to be missing image--calling redo()")
-            self.redo(channel, image)
+        self.panimage.clear()
 
     def reconfigure(self, panimage, width, height):
         self.logger.debug("new pan image dimensions are %dx%d" % (
@@ -270,40 +215,37 @@ class Pan(GingaPlugin.GlobalPlugin):
         panimage.redraw(whence=0)
         return True
 
-    def redraw_cb(self, fitsimage, whence, channel):
+    def redraw_cb(self, fitsimage, whence):
         if not self.gui_up:
             return
-        paninfo = channel.extdata._pan_info
-        if paninfo is not None:
-            if whence < 3:
-                fitsimage.copy_attributes(paninfo.panimage, self.copy_attrs,
-                                          whence=whence)
-                paninfo.panimage.zoom_fit()
-                self.panset(channel.fitsimage, channel, paninfo)
-            else:
-                # nothing except a graphical overlay change, so simply
-                # redraw to keep pace
-                paninfo.panimage.redraw(whence=whence)
+        if whence < 3:
+            self.fitsimage.copy_attributes(self.panimage, self.copy_attrs,
+                                           whence=whence)
+            self.panimage.zoom_fit()
+            self.panset()
+        else:
+            # nothing except a graphical overlay change, so simply
+            # redraw to keep pace
+            self.panimage.redraw(whence=whence)
         return True
 
     # LOGIC
 
     def clear(self):
-        if self.info is not None:
-            self.info.panimage.clear()
+        self.panimage.clear()
 
-    def set_image(self, channel, paninfo, image):
-        if image is None or not paninfo.panimage.viewable(image):
+    def set_image(self, image):
+        if image is None or not self.panimage.viewable(image):
             self.logger.debug("no main image--clearing Pan viewer")
-            paninfo.panimage.clear()
+            self.panimage.clear()
             return
 
-        paninfo.panimage.zoom_fit()
+        self.panimage.zoom_fit()
 
-        p_canvas = paninfo.panimage.get_private_canvas()
+        p_canvas = self.panimage.get_private_canvas()
         # remove old WCS compass
         try:
-            p_canvas.delete_object(paninfo.compass_wcs)
+            p_canvas.delete_object(self.compass_wcs)
         except Exception:
             pass
 
@@ -314,7 +256,7 @@ class Pan(GingaPlugin.GlobalPlugin):
                 # HACK: force a wcs error here if one is going to happen
                 wcs.add_offset_xy(image, x, y, 1.0, 1.0)
 
-                p_canvas.add_object(paninfo.compass_wcs)
+                p_canvas.add_object(self.compass_wcs)
             except Exception as e:
                 self.logger.warning("Can't calculate wcs compass: {}".format(e))
                 try:
@@ -326,17 +268,17 @@ class Pan(GingaPlugin.GlobalPlugin):
                     tb_str = "Traceback information unavailable."
                     self.logger.debug(tb_str)
 
-        self.panset(channel.fitsimage, channel, paninfo)
+        self.panset()
 
-    def panset(self, fitsimage, channel, paninfo):
-        image = fitsimage.get_image()
-        if image is None or not paninfo.panimage.viewable(image):
-            paninfo.panimage.clear()
+    def panset(self):
+        image = self.fitsimage.get_image()
+        if image is None or not self.panimage.viewable(image):
+            self.panimage.clear()
 
-        x, y = fitsimage.get_pan()
-        points = fitsimage.get_pan_bbox()
+        x, y = self.fitsimage.get_pan()
+        points = self.fitsimage.get_pan_bbox()
 
-        limits = fitsimage.get_limits()
+        limits = self.fitsimage.get_limits()
         width = limits[1][0] - limits[0][0]
         height = limits[1][1] - limits[0][1]
         edgew = math.sqrt(width**2 + height**2)
@@ -344,18 +286,18 @@ class Pan(GingaPlugin.GlobalPlugin):
 
         # Mark pan rectangle and pan position
         self.logger.debug("starting panset")
-        p_canvas = paninfo.panimage.get_private_canvas()
-        if paninfo.panrect is None:
-            paninfo.panrect = self.dc.CompoundObject(
+        p_canvas = self.panimage.get_private_canvas()
+        if self.panrect is None:
+            self.panrect = self.dc.CompoundObject(
                 self.dc.Point(
                     x, y, radius, style='plus',
                     color=self.settings.get('pan_position_color', 'yellow')),
                 self.dc.Polygon(
                     points,
                     color=self.settings.get('pan_rectangle_color', 'red')))
-            p_canvas.add_object(paninfo.panrect)
+            p_canvas.add_object(self.panrect)
         else:
-            point, bbox = paninfo.panrect.objects
+            point, bbox = self.panrect.objects
             point.x, point.y = x, y
             point.radius = radius
             bbox.points = points
@@ -365,27 +307,30 @@ class Pan(GingaPlugin.GlobalPlugin):
         #paninfo.panimage.zoom_fit()
         return True
 
+    def get_channel_viewer(self):
+        return self.channel.get_viewer('Ginga Image')
+
     def motion_cb(self, fitsimage, event, data_x, data_y):
-        chviewer = self.fv.getfocus_viewer()
+        chviewer = self.get_channel_viewer()
         self.fv.showxy(chviewer, data_x, data_y)
         return True
 
     def drag_cb(self, fitsimage, event, data_x, data_y):
         # this is a panning move in the small
         # window for the big window
-        chviewer = self.fv.getfocus_viewer()
+        chviewer = self.get_channel_viewer()
         chviewer.panset_xy(data_x, data_y)
         return True
 
     def btndown(self, fitsimage, event, data_x, data_y):
-        chviewer = self.fv.getfocus_viewer()
+        chviewer = self.get_channel_viewer()
         chviewer.panset_xy(data_x, data_y)
         return True
 
     def zoom_cb(self, fitsimage, event):
         """Zoom event in the pan window.  Just zoom the channel viewer.
         """
-        chviewer = self.fv.getfocus_viewer()
+        chviewer = self.get_channel_viewer()
         bd = chviewer.get_bindings()
 
         mode = bd.get_mode_obj('pan')
@@ -394,7 +339,7 @@ class Pan(GingaPlugin.GlobalPlugin):
     def zoom_pinch_cb(self, fitsimage, event):
         """Pinch event in the pan window.  Just zoom the channel viewer.
         """
-        chviewer = self.fv.getfocus_viewer()
+        chviewer = self.get_channel_viewer()
         bd = chviewer.get_bindings()
 
         mode = bd.get_mode_obj('pan')
@@ -403,7 +348,7 @@ class Pan(GingaPlugin.GlobalPlugin):
     def pan_pan_cb(self, fitsimage, event):
         """Pan event in the pan window.  Just pan the channel viewer.
         """
-        chviewer = self.fv.getfocus_viewer()
+        chviewer = self.get_channel_viewer()
         bd = chviewer.get_bindings()
 
         mode = bd.get_mode_obj('pan')
@@ -419,7 +364,7 @@ class Pan(GingaPlugin.GlobalPlugin):
             return False
         xc = (obj.x1 + obj.x2) / 2.0
         yc = (obj.y1 + obj.y2) / 2.0
-        chviewer = self.fv.getfocus_viewer()
+        chviewer = self.get_channel_viewer()
         # note: chviewer <-- referring to channel viewer
         with chviewer.suppress_redraw:
             chviewer.panset_xy(xc, yc)
@@ -444,7 +389,7 @@ class Pan(GingaPlugin.GlobalPlugin):
         return True
 
     def __str__(self):
-        return 'pan'
+        return 'panchild'
 
 
 # Append module docstring with config doc for auto insert by Sphinx.